+++ ArkanoidClone/Assets/Scripts/Player/MovementLeft.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 511d2d65bc72b974fa7a00ecb5fac571
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
