+++ ArkanoidClone/Assets/Scripts/Player/MovementRigth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eca3a994cbb1a484ba78a6fa2b04e00d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
